requirements.txt
Flask>=2.3.3
flask-cors>=4.0.0
Pillow>=10.0.0
numpy>=1.26.4
tensorflow>=2.16.1
Werkzeug>=2.3.7
